pyproject.toml
[tool.poetry]
name = "twitter-fake"
version = "0.1.0"
description = ""
authors = ["JaninaVolkmer <janina.volkmer@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.3"
djangorestframework = "^3.15.0"
pytest-django = "^4.8.0"
djangorestframework-simplejwt = "^5.3.1"
pillow = "^10.2.0"
drf-extra-fields = "^3.7.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "twitter_fake.settings"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
